+++ docs/source/NEWS.rst
@@ -19,6 +19,8 @@ Changes
 
 - Using modern classmethod decorator instead of old-style method call.
 - Usage of zope.interfaces was updated in preparation for python3 port.
+- ``toWire`` method is used to get bytes representation of ``ldaptor.protocols.pureber`` and
+  ``ldaptor.protocols.pureldap`` classes instead of ``__str__`` which is deprecated now.
 - Code was updated to pass `python3 -m compileall` in preparation for py3 port.
 - Continuous test are executed only against latest related Twisted and latest
   Twisted trunk branch.